+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Client:setNickname now uses proper endpoint
 - Fixed issue where nickname would not update
 - Fixed issue where deleting private channels crashed library
+- Added table.randompair and table.randomipair to stdlib ext
 
 
 ### 0.4.0